devices/analog-meter-readers/led-meter/requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
paho-mqtt==1.5.1
python-dotenv==0.19.0